# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Otsuka API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a new user / register",
                "description": "This endpoint allows you to create a new user / register",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role_id",
                                    "institute",
                                    "province_id",
                                    "regency_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "institute": {
                                        "type": "string",
                                        "example": "XYZ University"
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "regency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password1"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string",
                                                "property": null
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - User could not be created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "role_id": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 1
                    },
                    "role_name": {
                        "description": "Role",
                        "type": "string",
                        "example": "Role"
                    },
                    "email": {
                        "description": "User's email",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-02T10:00:00Z"
                    },
                    "deleted_at": {
                        "description": "User update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-02T14:20:00Z"
                    },
                    "created_at": {
                        "description": "User creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-01T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "User update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-02T14:20:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_auth": {
                "type": "http",
                "description": "Enter your JWT token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}